# chore(job): optimized memory allocation