--- a/job/job_test.go
+++ b/job/job_test.go
@@ -1,0 +1,1 @@
+package job