--- a/job/option.go
+++ b/job/option.go
@@ -8,54 +8,45 @@ type Options struct {
 	timeout    time.Duration
 }
 
-// An Option configures a mutex.
-type Option interface {
-	apply(*Options)
-}
-
-// OptionFunc is a function that configures a job.
-type OptionFunc func(*Options)
-
-// apply calls f(option)
-func (f OptionFunc) apply(option *Options) {
-	f(option)
-}
-
-func newDefaultOptions() *Options {
-	return &Options{
+func newDefaultOptions() Options {
+	return Options{
 		retryCount: 0,
 		retryDelay: 100 * time.Millisecond,
 		timeout:    60 * time.Minute,
 	}
 }
 
+type AllowOption struct {
+	RetryCount *int64
+	RetryDelay *time.Duration
+	Timeout    *time.Duration
+}
+
 // NewOptions with custom parameter
-func NewOptions(opts ...Option) *Options {
+func NewOptions(opts ...AllowOption) Options {
 	o := newDefaultOptions()
 
-	// Loop through each option
-	for _, opt := range opts {
-		// Call the option giving the instantiated
-		opt.apply(o)
+	if len(opts) != 0 {
+		if opts[0].RetryCount != nil && *opts[0].RetryCount != o.retryCount {
+			o.retryCount = *opts[0].RetryCount
+		}
+
+		if opts[0].RetryDelay != nil && *opts[0].RetryDelay != o.retryDelay {
+			o.retryDelay = *opts[0].RetryDelay
+		}
+
+		if opts[0].Timeout != nil && *opts[0].Timeout != o.timeout {
+			o.timeout = *opts[0].Timeout
+		}
 	}
 
 	return o
 }
 
-func WithRetryCount(count int64) Option {
-	return OptionFunc(func(o *Options) {
-		o.retryCount = count
-	})
+func Int64(val int64) *int64 {
+	return &val
 }
 
-func WithRetryDelay(t time.Duration) Option {
-	return OptionFunc(func(o *Options) {
-		o.retryDelay = t
-	})
-}
-
-func WithTimeout(t time.Duration) Option {
-	return OptionFunc(func(o *Options) {
-		o.timeout = t
-	})
+func Time(v time.Duration) *time.Duration {
+	return &v
 }