## old

```
goos: linux
goarch: amd64
pkg: github.com/golang-queue/queue
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
BenchmarkQueueTask
BenchmarkQueueTask-2         	 7688836	       137.1 ns/op	      99 B/op	       2 allocs/op
BenchmarkQueueTask-2         	 7391805	       161.6 ns/op	     101 B/op	       2 allocs/op
BenchmarkQueueTask-2         	 7356476	       180.4 ns/op	     101 B/op	       2 allocs/op
BenchmarkQueueTask-2         	 6143602	       201.7 ns/op	     101 B/op	       2 allocs/op
BenchmarkQueueTask-2         	 5050410	       201.0 ns/op	     102 B/op	       2 allocs/op
BenchmarkQueue
BenchmarkQueue-2             	 1777933	       747.7 ns/op	     290 B/op	       5 allocs/op
BenchmarkQueue-2             	 1415230	       794.5 ns/op	     294 B/op	       6 allocs/op
BenchmarkQueue-2             	 1659774	      1007 ns/op	     299 B/op	       6 allocs/op
BenchmarkQueue-2             	 1674228	       785.8 ns/op	     292 B/op	       6 allocs/op
BenchmarkQueue-2             	 1369141	      2739 ns/op	     290 B/op	       5 allocs/op
BenchmarkConsumerPayload
BenchmarkConsumerPayload-2   	  433586	      2892 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerPayload-2   	  555051	      2631 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerPayload-2   	  549049	      2700 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerPayload-2   	  522826	      2254 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerPayload-2   	  543097	      2167 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerTask
BenchmarkConsumerTask-2      	  590689	      1925 ns/op	     592 B/op	      10 allocs/op
BenchmarkConsumerTask-2      	  600896	      1917 ns/op	     592 B/op	      10 allocs/op
BenchmarkConsumerTask-2      	  548496	      1919 ns/op	     592 B/op	      10 allocs/op
BenchmarkConsumerTask-2      	  547576	      1947 ns/op	     592 B/op	      10 allocs/op
BenchmarkConsumerTask-2      	  595194	      1947 ns/op	     592 B/op	      10 allocs/op
BenchmarkNewTask
BenchmarkNewTask-2     	11069827	       107.8 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewTask-2     	10980966	       107.4 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewTask-2     	11049184	       107.8 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewTask-2     	11085979	       107.3 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewTask-2     	10908338	       107.6 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewOption
BenchmarkNewOption-2   	11046718	       107.9 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewOption-2   	10913220	       108.2 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewOption-2   	10924417	       111.5 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewOption-2   	11088232	       108.9 ns/op	      72 B/op	       4 allocs/op
BenchmarkNewOption-2   	11000745	       109.4 ns/op	      72 B/op	       4 allocs/op
```

## New

```
goos: linux
goarch: amd64
pkg: github.com/golang-queue/queue
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
BenchmarkQueueTask
BenchmarkQueueTask-2         	12296089	       104.9 ns/op	      72 B/op	       1 allocs/op
BenchmarkQueueTask-2         	10956156	       113.6 ns/op	      73 B/op	       1 allocs/op
BenchmarkQueueTask-2         	12407688	       132.6 ns/op	      72 B/op	       1 allocs/op
BenchmarkQueueTask-2         	12778882	       137.2 ns/op	      73 B/op	       1 allocs/op
BenchmarkQueueTask-2         	 9458818	       144.6 ns/op	      73 B/op	       1 allocs/op
BenchmarkQueue
BenchmarkQueue-2             	  493566	      7608 ns/op	     278 B/op	       5 allocs/op
BenchmarkQueue-2             	 1866391	       617.2 ns/op	     272 B/op	       5 allocs/op
BenchmarkQueue-2             	 1642077	       640.4 ns/op	     272 B/op	       5 allocs/op
BenchmarkQueue-2             	 1896705	       694.9 ns/op	     265 B/op	       4 allocs/op
BenchmarkQueue-2             	  500755	     17309 ns/op	     260 B/op	       4 allocs/op
BenchmarkConsumerPayload
BenchmarkConsumerPayload-2   	  478116	      2437 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerPayload-2   	  493172	      2491 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerPayload-2   	  556701	      2495 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerPayload-2   	  535084	      2203 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerPayload-2   	  508018	      2211 ns/op	     624 B/op	      11 allocs/op
BenchmarkConsumerTask
BenchmarkConsumerTask-2      	  573092	      1938 ns/op	     592 B/op	      10 allocs/op
BenchmarkConsumerTask-2      	  547263	      1948 ns/op	     592 B/op	      10 allocs/op
BenchmarkConsumerTask-2      	  576758	      1926 ns/op	     592 B/op	      10 allocs/op
BenchmarkConsumerTask-2      	  571465	      1950 ns/op	     592 B/op	      10 allocs/op
BenchmarkConsumerTask-2      	  575342	      1935 ns/op	     592 B/op	      10 allocs/op
BenchmarkNewTask
BenchmarkNewTask-2     	294638179	         4.062 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewTask-2     	295681586	         4.067 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewTask-2     	295283884	         4.058 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewTask-2     	293375212	         4.078 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewTask-2     	294573591	         4.043 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewOption
BenchmarkNewOption-2   	297229272	         4.048 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewOption-2   	297776034	         4.037 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewOption-2   	294032257	         4.046 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewOption-2   	296694416	         4.054 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewOption-2   	293353922	         4.039 ns/op	       0 B/op	       0 allocs/op
```

## Result

```
$ benchstat a.txt b.txt 
name               old time/op    new time/op    delta
QueueTask-2           176ns ±22%     127ns ±17%   -28.23%  (p=0.032 n=5+5)
Queue-2             1.21µs ±125%   5.37µs ±222%      ~     (p=0.690 n=5+5)
ConsumerPayload-2    2.53µs ±14%    2.37µs ± 7%      ~     (p=0.421 n=5+5)
ConsumerTask-2       1.93µs ± 1%    1.94µs ± 1%      ~     (p=0.198 n=5+5)
NewTask-2             108ns ± 0%       4ns ± 0%   -96.22%  (p=0.008 n=5+5)
NewOption-2           109ns ± 2%       4ns ± 0%   -96.30%  (p=0.008 n=5+5)

name               old alloc/op   new alloc/op   delta
QueueTask-2            101B ± 2%       73B ± 1%   -27.98%  (p=0.008 n=5+5)
Queue-2                293B ± 2%      269B ± 3%    -8.05%  (p=0.008 n=5+5)
ConsumerPayload-2      624B ± 0%      624B ± 0%      ~     (all equal)
ConsumerTask-2         592B ± 0%      592B ± 0%      ~     (all equal)
NewTask-2             72.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
NewOption-2           72.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
QueueTask-2            2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.008 n=5+5)
Queue-2                5.60 ±11%      4.60 ±13%      ~     (p=0.079 n=5+5)
ConsumerPayload-2      11.0 ± 0%      11.0 ± 0%      ~     (all equal)
ConsumerTask-2         10.0 ± 0%      10.0 ± 0%      ~     (all equal)
NewTask-2              4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
NewOption-2            4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```

Signed-off-by: Bo-Yi.Wu <appleboy.tw@gmail.com>